Add RepoAnalyzerFactory
