+++ analyzer/repo-analyzer/src/test/java/eu/fasten/analyzer/repoanalyzer/repo/GradleRepoAnalyzerTest.java
@@ -39,7 +39,8 @@ class GradleRepoAnalyzerTest {
     static void setUp() {
         root = new File(Objects.requireNonNull(GradleRepoAnalyzerTest.class.getClassLoader()
                 .getResource("simpleGradleRepo")).getFile()).getAbsolutePath();
-        analyzer = RepoAnalyzer.of(root);
+        var repoAnalyzerFactory = new RepoAnalyzerFactory();
+        analyzer = repoAnalyzerFactory.getAnalyzer(root);
     }
 
     @Test
@@ -68,6 +69,7 @@ class GradleRepoAnalyzerTest {
     void gradleKotlin() {
         root = new File(Objects.requireNonNull(GradleRepoAnalyzerTest.class.getClassLoader()
                 .getResource("simpleGradleKotlinRepo")).getFile()).getAbsolutePath();
-        analyzer = RepoAnalyzer.of(root);
+        var repoAnalyzerFactory = new RepoAnalyzerFactory();
+        analyzer = repoAnalyzerFactory.getAnalyzer(root);
     }
 }
